+++ backend/src/main/java/moaon/backend/article/repository/querymodifier/ClicksCursorModifier.java
@@ -5,8 +5,8 @@ import static moaon.backend.article.domain.QArticle.article;
 import com.querydsl.core.BooleanBuilder;
 import lombok.RequiredArgsConstructor;
 import moaon.backend.article.dto.ArticleQueryCondition;
-import moaon.backend.global.domain.QueryModifier;
 import moaon.backend.global.domain.cursor.Cursor;
+import moaon.backend.global.repository.QueryModifier;
 
 @RequiredArgsConstructor
 public class ClicksCursorModifier implements QueryModifier<Void, ArticleQueryCondition> {
