+++ backend/src/main/java/moaon/backend/article/repository/querymodifier/SearchKeywordModifier.java
@@ -9,8 +9,8 @@ import java.util.Arrays;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import moaon.backend.article.dto.ArticleQueryCondition;
-import moaon.backend.global.domain.QueryModifier;
 import moaon.backend.global.domain.SearchKeyword;
+import moaon.backend.global.repository.QueryModifier;
 import org.hibernate.boot.model.FunctionContributions;
 import org.hibernate.boot.model.FunctionContributor;
 import org.hibernate.query.sqm.function.SqmFunctionRegistry;
