+++ backend/src/main/java/moaon/backend/project/domain/cursor/LovesProjectCursor.java
@@ -1,17 +1,18 @@
-package moaon.backend.project.domain;
+package moaon.backend.project.domain.cursor;
 
 import moaon.backend.global.domain.cursor.AbstractCursor;
 import moaon.backend.global.domain.cursor.formatter.CursorFormatter;
 import moaon.backend.global.domain.cursor.formatter.IntegerFormatter;
+import moaon.backend.project.domain.Project;
 
-public class ViewsProjectCursor extends AbstractCursor<Integer> {
+public class LovesProjectCursor extends AbstractCursor<Integer> {
 
-    public ViewsProjectCursor(String rawCursor) {
+    public LovesProjectCursor(String rawCursor) {
         super(rawCursor);
     }
 
-    public ViewsProjectCursor(Project project) {
-        super(project.getViews(), project.getId());
+    public LovesProjectCursor(Project project) {
+        super(project.getLoveCount(), project.getId());
     }
 
     @Override
