+++ backend/src/main/java/moaon/backend/project/domain/cursor/ViewsProjectCursor.java
@@ -1,21 +1,18 @@
-package moaon.backend.article.domain;
+package moaon.backend.project.domain.cursor;
 
 import moaon.backend.global.domain.cursor.AbstractCursor;
 import moaon.backend.global.domain.cursor.formatter.CursorFormatter;
 import moaon.backend.global.domain.cursor.formatter.IntegerFormatter;
+import moaon.backend.project.domain.Project;
 
-public class ClickArticleCursor extends AbstractCursor<Integer> {
+public class ViewsProjectCursor extends AbstractCursor<Integer> {
 
-    public ClickArticleCursor(String rawCursor) {
+    public ViewsProjectCursor(String rawCursor) {
         super(rawCursor);
     }
 
-    public ClickArticleCursor(Article article) {
-        super(article.getClicks(), article.getId());
-    }
-
-    public ClickArticleCursor(Integer clicks, Long id) {
-        super(clicks, id);
+    public ViewsProjectCursor(Project project) {
+        super(project.getViews(), project.getId());
     }
 
     @Override
