+++ backend/src/main/java/moaon/backend/project/repository/querymodifier/CursorModifier.java
@@ -2,11 +2,11 @@ package moaon.backend.project.repository.querymodifier;
 
 import com.querydsl.core.BooleanBuilder;
 import lombok.RequiredArgsConstructor;
-import moaon.backend.global.domain.QueryModifier;
 import moaon.backend.global.domain.cursor.Cursor;
-import moaon.backend.project.domain.CreatedAtCursor;
-import moaon.backend.project.domain.LovesProjectCursor;
-import moaon.backend.project.domain.ViewsProjectCursor;
+import moaon.backend.global.repository.QueryModifier;
+import moaon.backend.project.domain.cursor.CreatedAtCursor;
+import moaon.backend.project.domain.cursor.LovesProjectCursor;
+import moaon.backend.project.domain.cursor.ViewsProjectCursor;
 import moaon.backend.project.dto.ProjectQueryCondition;
 
 @RequiredArgsConstructor
