+++ backend/src/main/java/moaon/backend/project/repository/querymodifier/LovesCursorModifier.java
@@ -4,8 +4,8 @@ import static moaon.backend.project.domain.QProject.project;
 
 import com.querydsl.core.BooleanBuilder;
 import lombok.RequiredArgsConstructor;
-import moaon.backend.global.domain.QueryModifier;
 import moaon.backend.global.domain.cursor.Cursor;
+import moaon.backend.global.repository.QueryModifier;
 import moaon.backend.project.dto.ProjectQueryCondition;
 
 @RequiredArgsConstructor
