+++ backend/src/main/java/moaon/backend/project/repository/querymodifier/SearchKeywordModifier.java
@@ -8,8 +8,8 @@ import com.querydsl.core.types.dsl.Expressions;
 import java.util.Arrays;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
-import moaon.backend.global.domain.QueryModifier;
 import moaon.backend.global.domain.SearchKeyword;
+import moaon.backend.global.repository.QueryModifier;
 import moaon.backend.project.dto.ProjectQueryCondition;
 import org.hibernate.boot.model.FunctionContributions;
 import org.hibernate.boot.model.FunctionContributor;
