+++ backend/src/main/java/moaon/backend/project/repository/querymodifier/SortByModifier.java
@@ -3,7 +3,7 @@ package moaon.backend.project.repository.querymodifier;
 import static moaon.backend.project.domain.QProject.project;
 
 import com.querydsl.core.types.OrderSpecifier;
-import moaon.backend.global.domain.QueryModifier;
+import moaon.backend.global.repository.QueryModifier;
 import moaon.backend.project.domain.ProjectSortType;
 import moaon.backend.project.dto.ProjectQueryCondition;
