+++ backend/src/test/java/moaon/backend/article/dto/CreatedAtArticleCursorTest.java
@@ -3,7 +3,7 @@ package moaon.backend.article.dto;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDateTime;
-import moaon.backend.article.domain.CreatedAtArticleCursor;
+import moaon.backend.article.domain.cursor.CreatedAtArticleCursor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
