+++ backend/src/test/java/moaon/backend/article/repository/CustomizedArticleRepositoryImplTest.java
@@ -7,8 +7,8 @@ import java.util.List;
 import moaon.backend.article.domain.Article;
 import moaon.backend.article.domain.ArticleCategory;
 import moaon.backend.article.domain.ArticleSortType;
-import moaon.backend.article.domain.ClickArticleCursor;
-import moaon.backend.article.domain.CreatedAtArticleCursor;
+import moaon.backend.article.domain.cursor.ClickArticleCursor;
+import moaon.backend.article.domain.cursor.CreatedAtArticleCursor;
 import moaon.backend.article.dto.ArticleQueryCondition;
 import moaon.backend.fixture.ArticleFixtureBuilder;
 import moaon.backend.fixture.ArticleQueryConditionBuilder;
